requirements.txt
mysql_connector_repackaged==0.3.1
opencv_python==4.6.0.66
pandas==1.4.3
pytesseract==0.3.9
selenium==4.3.0
